# cursor_sh/api-spec/openapi.yaml
openapi: 3.0.3
info:
  title: AI设计任务管理系统 API
  description: |
    VR+AI 裸眼3D内容定制管理系统的后端 API 接口规范
    
    ## 功能模块
    - **认证模块**: 用户登录、注册、登出
    - **订单模块**: 订单创建、查询、状态管理、反馈
    - **负责人模块**: 负责人管理、订单分配
    
    ## 订单类型
    1. **video_purchase** - 裸眼3D成片购买适配
    2. **ai_3d_custom** - AI裸眼3D内容定制
    3. **digital_art** - 数字艺术内容定制
    
    ## 订单状态流转
    ```
    pending_assign (待分配)
      ↓
    in_production (制作中)
      ↓
    preview_ready (初稿预览)
      ↓ (客户反馈)
    ├─ revision_needed (需要修改) → in_production
    └─ in_production → final_preview (终稿预览) → completed (已完成)
    ```
  version: 1.0.0
  contact:
    name: API Support
    email: support@example.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000/api
    description: 本地开发环境
  - url: https://api.example.com/api
    description: 生产环境
...
security:
  - BearerAuth: []
paths:
...
  /staff:
    get:
      tags:
        - 负责人
      summary: 获取负责人列表
      description: 获取所有负责人（staff和admin角色）用户列表，支持分页和筛选
      parameters:
        - name: page
          in: query
          description: 页码
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: pageSize
          in: query
          description: 每页数量
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: keyword
          in: query
          description: 搜索关键词（搜索用户名、姓名、邮箱）
          schema:
            type: string
        - name: role
          in: query
          description: 角色筛选
          schema:
            type: string
            enum: [admin, staff]
        - name: isActive
          in: query
          description: 状态筛选
          schema:
            type: boolean
      responses:
        '200':
          description: 成功获取负责人列表
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          data:
                            type: array
                            items:
                              $ref: '#/components/schemas/StaffUser'
                          total:
                            type: integer
                            description: 总记录数
              example:
                code: 200
                message: "获取成功"
                data:
                  data:
                    - id: "staff-001"
                      username: "staff1"
                      role: "staff"
                      email: "staff1@example.com"
                      realName: "张设计"
                      isActive: true
                      orderCount: 5
                      createdAt: "2024-10-01T08:00:00Z"
                    - id: "staff-002"
                      username: "staff2"
                      role: "staff"
                      email: "staff2@example.com"
                      realName: "李艺术"
                      isActive: true
                      orderCount: 3
                      createdAt: "2024-10-15T09:30:00Z"
                  total: 2

    post:
      tags:
        - 负责人
      summary: 添加负责人
      description: 创建新的负责人账户（仅管理员可操作）
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StaffCreate'
            example:
              username: "staff3"
              password: "123456"
              email: "staff3@example.com"
              realName: "王制作"
              role: "staff"
              isActive: true
      responses:
        '201':
          description: 负责人创建成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/StaffUser'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          description: 用户名已存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              example:
                code: 409
                message: "用户名已存在"
                data: null
  /staff/{staffId}:
    put:
      tags:
        - 负责人
      summary: 更新负责人信息
      description: 更新负责人的信息（仅管理员可操作）
      parameters:
        - name: staffId
          in: path
          required: true
          schema:
            type: string
          description: 负责人ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StaffUpdate'
            example:
              email: "staff1_new@example.com"
              realName: "张设计师"
              role: "staff"
              isActive: true
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/StaffUser'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          description: 负责人不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              example:
                code: 404
                message: "负责人不存在"
                data: null

    delete:
      tags:
        - 负责人
      summary: 删除负责人
      description: 删除负责人账户（仅管理员可操作，且负责人不能有进行中的订单）
      parameters:
        - name: staffId
          in: path
          required: true
          schema:
            type: string
          description: 负责人ID
      responses:
        '200':
          description: 删除成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              example:
                code: 200
                message: "删除成功"
                data: null
        '400':
          description: 无法删除（有进行中的订单）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              example:
                code: 400
                message: "该负责人还有 5 个进行中的订单，无法删除"
                data: null
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          description: 负责人不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              example:
                code: 404
                message: "负责人不存在"
                data: null
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT Token 认证，格式：Bearer {token}
  schemas:
...
    StaffUser:
      type: object
      required:
        - id
        - username
        - role
        - isActive
      properties:
        id:
          type: string
          description: 负责人ID
          example: "staff-001"
        username:
          type: string
          description: 用户名
          example: "staff1"
        email:
          type: string
          format: email
          description: 邮箱
          example: "staff1@example.com"
        realName:
          type: string
          description: 真实姓名
          example: "张设计"
        role:
          type: string
          enum: [admin, staff]
          description: 角色（管理员或负责人）
          example: "staff"
        isActive:
          type: boolean
          description: 是否启用
          example: true
        orderCount:
          type: integer
          description: 负责的订单数量（进行中的订单）
          example: 5
        createdAt:
          type: string
          format: date-time
          description: 创建时间
          example: "2024-10-01T08:00:00Z"
        updatedAt:
          type: string
          format: date-time
          description: 更新时间
          example: "2024-11-01T10:30:00Z"
    StaffCreate:
      type: object
      required:
        - username
        - password
        - email
        - realName
        - role
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 50
          pattern: '^[a-zA-Z0-9_]+$'
          description: 用户名（只能包含字母、数字和下划线）
          example: "staff3"
        password:
          type: string
          minLength: 6
          format: password
          description: 密码
          example: "123456"
        email:
          type: string
          format: email
          description: 邮箱
          example: "staff3@example.com"
        realName:
          type: string
          description: 真实姓名
          example: "王制作"
        role:
          type: string
          enum: [admin, staff]
          description: 角色
          default: "staff"
          example: "staff"
        isActive:
          type: boolean
          description: 是否启用
          default: true
          example: true
    StaffUpdate:
      type: object
      properties:
        email:
          type: string
          format: email
          description: 邮箱
          example: "staff1_new@example.com"
        realName:
          type: string
          description: 真实姓名
          example: "张设计师"
        role:
          type: string
          enum: [admin, staff]
          description: 角色
          example: "staff"
        isActive:
          type: boolean
          description: 是否启用
          example: true

    # ========== 订单相关 ==========
...
    ApiResponse:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: integer
          description: 状态码
          example: 200
        message:
          type: string
          description: 响应消息
          example: "操作成功"
        data:
          description: 响应数据
          nullable: true
  responses:
    BadRequest:
      description: 请求参数错误
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiResponse'
          example:
            code: 400
            message: "请求参数错误"
            data: null
...
    Forbidden:
      description: 禁止访问，权限不足
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiResponse'
          example:
            code: 403
            message: "权限不足"
            data: null
...